docs(angular-query): remove msw from angular example
interceptor works better in sandbox environments
